fix: handle line breaks
