+++ app/utils/rich-text-render-options.tsx
@@ -1,0 +1,77 @@
+import { type Options } from "@contentful/rich-text-react-renderer"
+import {
+  BLOCKS,
+  type Block,
+  INLINES,
+  type Inline,
+} from "@contentful/rich-text-types"
+import { type ReactNode } from "react"
+
+export const defaultRichTextRenderOptions: Options = {
+  renderNode: {
+    [INLINES.HYPERLINK]: (node: Block | Inline, children: ReactNode) => {
+      const { data } = node
+      const { uri } = data
+      return (
+        <a
+          className="text-primary underline dark:text-gray-400"
+          target="_blank"
+          rel="noreferrer"
+          href={uri}
+        >
+          {children}
+        </a>
+      )
+    },
+    [BLOCKS.PARAGRAPH]: (node: Block | Inline, children: ReactNode) => {
+      return <p className="mb-4 text-base leading-relaxed">{children}</p>
+    },
+    [BLOCKS.HEADING_2]: (node: Block | Inline, children: ReactNode) => {
+      return <h2 className="mb-5 text-3xl">{children}</h2>
+    },
+    [BLOCKS.HEADING_3]: (node: Block | Inline, children: ReactNode) => {
+      return (
+        <h3 className="mb-4 text-2xl font-bold dark:text-gray-200">
+          {children}
+        </h3>
+      )
+    },
+    [BLOCKS.HEADING_4]: (node: Block | Inline, children: ReactNode) => {
+      return (
+        <h4 className="mb-4 text-xl uppercase dark:text-gray-200">
+          {children}
+        </h4>
+      )
+    },
+    [BLOCKS.HEADING_5]: (node: Block | Inline, children: ReactNode) => {
+      return (
+        <h5 className="mb-4 text-lg font-bold dark:text-gray-200">
+          {children}
+        </h5>
+      )
+    },
+    [BLOCKS.HEADING_6]: (node: Block | Inline, children: ReactNode) => {
+      return (
+        <h6 className="text-md mb-4 font-bold uppercase dark:text-gray-200">
+          {children}
+        </h6>
+      )
+    },
+    [BLOCKS.UL_LIST]: (node: Block | Inline, children: ReactNode) => {
+      return <ul className="ml-12 list-disc">{children}</ul>
+    },
+    [BLOCKS.OL_LIST]: (node: Block | Inline, children: ReactNode) => {
+      return <ol className="ml-12 list-decimal">{children}</ol>
+    },
+    [BLOCKS.LIST_ITEM]: (node: Block | Inline, children: ReactNode) => {
+      return <li>{children}</li>
+    },
+  },
+  renderText: (text: string) => {
+    return text
+      .split("\n")
+      .reduce<Array<ReactNode>>((children, textSegment, index) => {
+        return [...children, index > 0 && <br key={index} />, textSegment]
+      }, [])
+  },
+}
